Change KakaoOptionRequest, MessageParameter location
