+++ src/requests/messageRequest.ts
@@ -1,6 +1,8 @@
-import {Message} from '../models/message';
+import {Message, MessageType} from '../models/message';
 import {DateOperatorType} from '../types/commonTypes';
 import {formatWithTransfer} from '../lib/stringDateTrasnfer';
+import {kakaoOptionRequest} from './kakao/kakaoOptionRequest';
+import {RcsOptionRequest} from '../models/rcs/rcsOption';
 
 export type DefaultAgentType = {
   sdkVersion: string;
@@ -25,6 +27,22 @@ abstract class DefaultMessageRequest {
     this.allowDuplicates = false;
   }
 }
+
+export type MessageParameter = {
+  to: string | Array<string>;
+  from?: string;
+  text?: string;
+  imageId?: string;
+  type?: MessageType;
+  subject?: string;
+  autoTypeDetect?: boolean;
+  kakaoOptions?: kakaoOptionRequest;
+  rcsOptions?: RcsOptionRequest;
+  country?: string;
+  customFields?: Record<string, string>;
+  replacements?: Array<object>;
+  faxOptions?: FileIds;
+};
 
 export class SingleMessageSendingRequest extends DefaultMessageRequest {
   message: Message;
